Updated App.js
Made it so it is a 10 question quiz instead of infinite because I could not figure out how to do that with out current code structure. I added random questions with proper answer and random other choices. I also added a restart button as well as a short description to the page.